Directory reorganisation
This reorganisation aims to place the
SOS client code in the master/sos/
subdirectory. The HTTP server code
will be in in master/http/. And the
CoAP server code (if any) will be in
master/coap/.

The master/ directory will contain
the glue code (configuration file read,
daemon part, cache control, etc.).