Use C++ shared_ptr type for messages
The use of C++ smart pointers allows for flexibility since we
no longer bother for message removal. Once the last reference
to a message is removed, the message is automatically removed.
This way, we can safely remove an old message from a list
(for example the list of sent messages, sos::sos::mlist_) and
continue to use it.