Fix a long standing bug
To build the outgoing message (e.g. for the Assoc Answer message which
returns a .well-known text), one must know the available size in this
message. But, available size depends on the MTU which depends on the
L2 which is not defined at the time the message is built (it was given
as a parameter to send). So, define the message constructor as requiring
the L2.