Slave: remove a needless transition
When the slave is being associated with a Assoc message
from the master, without having received a Hello message,
its transition is :
	SL_WAITING_UNKNOWN -> SL_RUNNING
but the Hello id is still -1. When the Hello message is
received (with a valid Hello id), don't go to SL_WAITING_KNOWN
(and don't send a Discover request) before entering SL_RUNNING
again upon reception of another Assoc message.